usei, unit tests
Unit tests of usei to reproduce time-related issues.
I tried to adapt a POC usei network test, but the network tests are blackbox tests,
and we cannot create routing_manager_stub or
endpoint_manager_impl. Changing the interface was considered,
but it would not pass review. Another solution was to use debug
symbols to find the private methods, but that would be too
much. For these reasons, unit tests are developed.