fix race condition in connections_ map changes
In local communication, if a subscribe is sent to the service-provider,
but right before the subscription ack/nack is sent, the connection
service-provider -> service-consumer is removed, service-provider
detects and executes handle_client_error, however service-consumer
will not detect any connection problem and still be waiting for subscribe-ack.
What was happening is that right after replacing the old connection,
remove_connection was executed from ltsei::receive_cbk when detecting
an issue. This would then make the disconnect_from impossible to be executed
on the endpoint to the client, as no connection was present. The client would be
unaffected and not detect any problem keeping it in a "limbo" state.
By specifying the connection to remove, we can avoid removing a just-updated
connection and prevent this behavior.
Also made the connection's stop and destructor be executed outside of
connections_mutex_.