Introduce a thread-safe timer utility
Introduce a thread-safe timer utility
This timer utility can be used to either run a job periodically, or as a one shot timer.
This utility is going to be used in upcoming work.
Meanwhile introduce the capability to tinker with these timers,
e.g. in fake_socket tests.
Note that this would first require the actual use of this utility.