app: fail early, fail fast for unknown messages
If a user forgets to set the message type (so neither request, nor
response, nor notification...), the message will fail to arrive
anywhere, but often with a very cryptic error, such as:

rmi::send: Routing error. Endpoint for (...) could not be found!

This is too difficult to debug. Made it obvious!