Fix remote subscription expiration
This change fixes an edge-case in the subscription mechanism
where the routing host was able to send multiple unsubscribe
requests to the service for the same subscription.
Additionally, it also improves the logging of subscription requests
between applications.
Fix
This edge-case is visible when a remote client is using inconsistent
Session IDs to subscribe to the service, triggering multiple reboot
detections on the host. The host then sends multiple unsubscribe
requests while the service is still busy processing the first one.
The fix involves simply checking if a subscription has already been
expired before sending another unsubscribe requests to the service.
Log
The change to the logs adds the subscription id to the subscription
related traces, such SUBSCRIBE, UNSUBSCRIBE, SUBSCRIBE_ACK etc.
This makes it easier to track the subscription as it bounces between
service, client and host.
For example:
SUBSCRIBE(0000): [0001.0001.0001:ffff:0] accepted. id=0002 subscribers=2
SUBSCRIBE ACK(0101): [0001.0001.0001.ffff] id=0002