Deliver service availability only after endpoint creation
A possible race condition between the transition of a binded socket towards a connection oriented
UDP socket enables a scenario where packets can be dropped and a ICMP port unreachable sent back.
When the routing host processes an offer that is to be offered/subscribed exclusively on an UDP port,
the client endpoint is created, socket binded to local address and then it asynchronously starts the
connection oriented process (define remote address for endpoint), while doing it asynchronously, the
routing host sends the service availability to clients and might even receive their subscriptions and
send it to the remote partner via service discovery. This asynchronous behavior is considered to be
a possible root cause for the ICMP port unreachable message seen for the initial event being dropped
in
Other possible root causes for port unreachable are:

Port unavailable, socket not binded, do not believe is the scenario as the endpoint creation is
synchronous and previous to the availability.
Sender is not the remote address defined for the connection, also not the scenario here.

A fix is presented in this PR where the service availability is only sent to the clients when the
UDP sockets is considered "connected".
Climate test is updated as availability handler was called more than the expected number of times
as the offer was processed with the service not being requested but since an availability handler
was still registered and the client is the host, the handler was invoked with the
service state: AS_OFFERED.