tests: fix offer_test_external
It is a somewhat psychotic test, where first host offers a service, and
a second host offers the same service again, to check that collisions do
not break communication for the first offer
That is fine, however the second host would.. trigger a shutdown
method.. through the service that has a collision? Which makes no sense,
and was not deterministic. It is a leftover from b69aac5
Instead, have the high-level test logic stop the second host