Reset sending flag for UDP server endpoint
Reset is sending flag for UDP server endpoint.
An issue has been identified after STR where the the unicast socket for SD stopped
sending data to only one of its targets. Moreover, the dangerous ESQ trace that
depicts when a sending queue surpasses the warning limit was being logged
for that endpoint, this indicated that the data for that target reached the internal queue
but never the network, leading to believe that could be related with the is_sending flag
not being reset if an asynchronous operation was somehow halted/corrupted during STR.
This PR introduces some further traces to track targets unable to send data for the SD
endpoint and resets the is_sending_ flag for all UDP server endpoint on resume.