check config for app name and id
Check config for app name and id
If an app with clientID X and name Y de-registers and another Y name
app comes up and takes its clientID, if the original app re-registers
again, routing manager will accept the connection believing it's the
exact same application. This will then lead to issues where the second
app will detect lost of connection due to the replacing of connection
and try to register again, leading to a loop of register->de-register
between the two apps that share the same clientID X and name Y.
To fix this, check the config for configured app names, if it exists,
then it's OK, otherwise, continue with the clientID assignment
Note: there are several apps that do not specify a name, so it will
default to "" which can lead to the issue described above.