Optimize external_local_payload_test_client_external
Optimize external_local_payload_test_client_external
The test starter uses a function to verify if the service sockets are open, but unlike other starters, it checks after
launching the client instead of checking after launching the service.
Additionally, to verify if the sockets are open, a 2-second sleep is performed, but I don't think it makes sense to
always wait 2 seconds because in the case of sanitizers, it is almost instantaneous.
Therefore, to optimize execution times, a while loop was created that checks the socket status every 250ms with a
maximum limit of 2 seconds. This way, as soon as the sockets are open, the test can start, avoiding the static 2-second
wait.