workaround to race condition with restart list
Eliminate the risk of desynchronisation of the list of servers to be restarted
Since 4ef740c,
the lock on endpoint_mutex_ isn't hold while restarting the server.
So there is a risk that the list of active servers is modified while
the endpoint_manager_impl restarts them during a resume.
The consequence is that the old server occupies an endpoint
and prevents its replacement from taking its place.
We can reduce the likelihood of such an event occurring. And
when it occurs, we can correct the consequences.
Secondly, this problem should not occur because it is not normal
for the old server not to have self-destructed, as it should no longer
be referenced.