dissector: remove initial/end tag check
Remove initial/end tag check
With 7b3dedc7fcee, start and end tags were removed so vsomeip-dissector
had to be updated to no longer count on this 8 bytes.
Note: this breaks compatibility with pcaps that had versions previous
to the introduction of the PR mentioned above, you can use the
official wireshark vsomeip dissector for those.