Ensure buffer is not cleaned up while asio might use it
Ensure memory keeps alive until callback is invoked
keep a shared_ptr of the local_receive_buffer and copy it into the callback.
Because the callback does not keep alive the owner of the buffer as in the other
places, the callback needs to ensure that at least the memory used by asio is
kept alive. This avoids circular ownership problems, while protecting against
memory corruption.