Ensure lazy load of client, before using security
Ensure lazy load before using security
During the refactoring in: 7b3dedc7fcee
the add_known_client was moved after the creation of a local_endpoint.
But the creation method of the local_endpoint internally uses the security model to ensure
this endpoint is allowed to be used. This was done because:

It was not understood that add_known_client needs to be called strictly before the check,
there was no remove_client method in the rh interface to remove the client
when the security check failed.

Both of these points are now understood and adjusted in this PR.
This bug did not surface, because the lazy load + security check needs to be used for
a client-server relation across containers.