Avoid security failures on the daemon side
Avoid wrong security checks on the sender endpoints of the router and harden the local_server
Sequence that revealed all these problems addressed:

A container client that did not load the security policy (lazy loaded) tried to connect to the router
The router accepted the connection and assigned a client id
The client received the client id
The client started the receiver (triggered a async_accept)
The client dispatched the register_application command + the config command
The router enqueued the registration answer process on some other thread
The router started to load the security policy for the received config command
The routers endpoint was able to connect to the client
The router executed a "client" security check which failed
The router dropped the sender towards the client
The clients error handler is invoked
The client invoked rmc::reconnect but a credentials check within fails
The client does not stop the receiver
The register_application timer expires, not stopping the receiver, but restarting the sender
The loop starts from the beginning

Eventually the router is no longer rejecting the client (the policy loading was quick enough),
but now the pilled up async_accepts fail with "Bad file descriptor".
Problems address in this PR:

The sender of the router should not perform the security check (be aware it is a router endpoint)
The local_server should be able to handle subsequent start, start invocations,
Sockets used in the async_accept callback should not be replaced, before the callback is invoked.
remove the credential check when re-attempting to connect to the router