Log event caching
This change adds an explicit log when an event payload is set or
unset by its provider.
The log will help debug issues when it's unclear if an event
payload has been set or not by the application.
This matters because vsomeip will only trigger notifications if
the payload has been set.