avoid is-allowed call as sender
Avoid any security check on the sender, but create mapping in the router.
Before the refactoring of:
fa13022d00a4
The very first message received would have been a special one for uds
that contained the host name of the client. With this information
the router could lazy_load the security policy when the assign_client
command was received for this client and also invoke check_credentials.
The later does not only perform a check, it also builds up a mapping that
the router would later distribute.
To replicate this behavior the following changes are done in this PR:

config_id is always the first message to transmit the host name (also to the router)
the router stores the host name and then waits for assign client
the sender does not call check_credentials (don't invoke is_allowed)

On top:
In case a client had a wrong security setup before the mentioned refactoring:
The router would have not send an assign client id command out, but instead rejected
the connection early on. On the client side this lead to a reconnection attempt
that was interrupted due to a credentials check.
After the mentioned refactoring:

Router assigns the client id, drops the connections and "leaks" the id
Client busy loops reconnecting

To circumvent the leak, the id is reclaimed upon rejecting the connection,
to circumvent the busy loop a debounce of the restart of the sender is introduced.
The former credentials check on reconnect is problematic, as the security policy
might not be loaded yet.
Because this change implies that there is no sender even though rmc is started,
logs were added highlighting that something was not send, due to the missing
sender.