Enable dropping and injection of vsomeip commands
Enable dropping and injection of vsomeip commands.
This PR introduces the ability to drop and inject specific vsomeip commands to test following scenarios:

Clients behavior when an expected message is not delivered;
Deliver unexpected commands to tests application robustness;
Swap content of vsomeip commands;

This is achieved by registering a handler that is invoked for each parsed commands to inquire on
delivery decision.