make reuse_clientID_test more robust
Make reuse_clientID_test more robust
Due to valgrind being slow, it could happen that, due to the sleep between
app1 stop and start, app5 would begin but not register in time, allowing
app1 to re-register with the same clientID and app5 to register with a
different clientID than the one expected (app1's original clientID).
So changed the app1 restart to manually trigger when app5 registers.
Other tiny changes:

Increased the time that app1 has to re-register
Decrease the timeout of the app that does not register
Fixed test name