Fix availability inconsistency
Fix + tests for availability inconsistency
It was noticed that for a client only requesting a service an
ON_UNAVAILABLE call was missing, due to which a second
ON_AVAILABLE call was not propagated to the application.
This was happening because upon server deregistration, the daemon was clearing
the information about the offered services, before sending information about
the unavailability of the services to other connected clients. Thus, the ON_UNAVAILABLE
would not be propagated to requesters of the leaving services.
NOTE: This happens in the case of an 'ungraceful' shutdown, where the server
does not explicitly send STOP_OFFERS before shutting down